% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{asdm}
\alias{asdm}
\title{Approximate Species Distribution Model}
\usage{
asdm(
  xylist,
  map1,
  map2,
  map3,
  valuescale = 1000,
  overlap,
  k1 = 1,
  k2 = 1,
  datasplit = c(1, 0, 0),
  k_rep = 10,
  tss_rep = 10,
  ACC_rep = 10,
  REC_rep = 10,
  F1_rep = 10,
  valiMaxTech = "entropy",
  dJ = 1,
  kE.p = 999,
  step = 0.05,
  threshold = 0.001,
  k1.train = T,
  k2.train = T
)
}
\arguments{
\item{xylist}{Thinned coordinates of the target species, dataframe with two columns (X, Y)}

\item{map1}{Predicted distribution map of species A (Raster* classes, one layer)}

\item{map2}{Predicted distribution map of species B (Raster* classes, one layer)}

\item{map3}{Predicted distribution map of combination data (Raster* classes, one layer)}

\item{valuescale}{The theoretical maximum of inputted predicted distribution map (1 or 1000)}

\item{overlap}{The overlap value matrix/dataframe(3*3) of target species, A and B (follow this order of species)}

\item{k1}{Initial value of k1 (a float or "random" which would create a float between 0.7 and 1.3)}

\item{k2}{Initial value of k2 (a float or "random" which would create a float between 0.7 and 1.3)}

\item{datasplit}{Vector with 3 numbers in range of 0-1, which referred to: the proportion of train, validation and test data}

\item{k_rep}{Replicate number of k calculation}

\item{tss_rep}{Replicate number of sampling when calculating maxtss value}

\item{ACC_rep}{Replicate number of sampling when calculating ACC value}

\item{REC_rep}{Replicate number of sampling when calculating REC value}

\item{F1_rep}{Replicate number of sampling when calculating F1 value}

\item{valiMaxTech}{Maximized criterion by validation, only support "entropy" and "maxtss"}

\item{dJ}{Gradient}

\item{kE.p}{Initial kE}

\item{step}{Step}

\item{threshold}{Exit threshold}

\item{k1.train}{Whether k1 is trainable, TRUE or FALSE}

\item{k2.train}{Whether k2 is trainable, TRUE or FALSE}
}
\description{
Developing Approximate Species Distribution Model.
}
