% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.entropy}
\alias{calcu.map.entropy}
\title{Calculate entropy}
\usage{
calcu.map.entropy(map, xylist)
}
\arguments{
\item{map}{The predicted distribution map of the target species (Raster* class)}

\item{xylist}{Coordinates of the target species (dataframe), with two columns (X and Y)}
}
\description{
Calculate entropy of model.
}
\examples{
# example code

}
