% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.F1}
\alias{calcu.map.F1}
\title{Calculate F1 Score}
\usage{
calcu.map.F1(xylist, map, rep = 1, threshold)
}
\arguments{
\item{xylist}{Coordinates (dataframe), with two columns (X and Y)}

\item{map}{A Raster* class, the output of ASDM}

\item{rep}{Replications}

\item{threshold}{Threshold of suitability}
}
\description{
Calculate F1 score of model.
}
\examples{
# example code

}
