% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.MAE}
\alias{calcu.map.MAE}
\title{Calculate MAE}
\usage{
calcu.map.MAE(map1, map2, cellnum)
}
\arguments{
\item{map1}{Map for evaluation (ASDM output)}

\item{map2}{Map for reference ('true' SDM output)}

\item{cellnum}{Total number of grids.}
}
\description{
Calculate MAE of model.
}
\examples{
# example code

}
