% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.meva}
\alias{calcu.map.meva}
\title{Calculate Model Evaluation for a Given Threshold Value}
\usage{
calcu.map.meva(xylist, map, threshold)
}
\arguments{
\item{xylist}{Coordinates (dataframe), with two columns (X and Y)}

\item{map}{A Raster* class, the output of ASDM}

\item{threshold}{Threshold of suitability, a float between 0 and 1}
}
\description{
Calculate Model Evaluation for a Given Threshold Value
}
\examples{
# example code

}
