% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.R2}
\alias{calcu.map.R2}
\title{Calculate R2}
\usage{
calcu.map.R2(map1, map2)
}
\arguments{
\item{map1}{Map for evaluation (ASDM output)}

\item{map2}{Map for reference ('true' SDM output)}
}
\description{
Calculate R2 of model.
}
