% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.REC}
\alias{calcu.map.REC}
\title{Calculate Recall Rate}
\usage{
calcu.map.REC(xylist, map, rep = 1, threshold)
}
\arguments{
\item{xylist}{Coordinates (dataframe), with two columns (X and Y)}

\item{map}{A Raster* class, the output of ASDM}

\item{rep}{Replications}

\item{threshold}{Threshold of suitability}
}
\description{
Calculate recall rate of model.
}
\examples{
# example code

}
