% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{calcu.map.RMSE}
\alias{calcu.map.RMSE}
\title{Calculate RMSE}
\usage{
calcu.map.RMSE(map1, map2, cellnum)
}
\arguments{
\item{map1}{Map for evaluation (ASDM output)}

\item{map2}{Map for reference ('true' SDM output)}

\item{cellnum}{Total number of grids.}
}
\description{
Calculate RMSE of model.
}
\examples{
# example code

}
