% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{getRepCalcuMAXTSS}
\alias{getRepCalcuMAXTSS}
\title{Calculate MaxTSS (easy way)}
\usage{
getRepCalcuMAXTSS(xylist, map, tss_rep)
}
\arguments{
\item{xylist}{Coordinates of the target species, dataframe with 2 columns(X and Y)}

\item{map}{The predicted distribution map of the target species (Raster* class)}

\item{tss_rep}{Replication time for calculating MaxTSS}
}
\description{
Calculate MaxTSS of model by a high-level approach.
}
\examples{
# example code

}
