% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdm.R
\name{repCalcuMAXTSS}
\alias{repCalcuMAXTSS}
\title{Calculate MaxTSS}
\usage{
repCalcuMAXTSS(xylist, map, rep = 1)
}
\arguments{
\item{xylist}{Coordinates (dataframe), with two columns (X and Y)}

\item{map}{A Raster* class, the output of ASDM}

\item{rep}{Replications}
}
\description{
Calculate MaxTSS of model.
}
